third_party/WebKit/LayoutTests/virtual/origin-trials-runtimeflags-disabled/http/tests/origin_trials/webexposed/imagecapture-origin-trial-interfaces-expected.txt

CONSOLE MESSAGE: line 17: Interfaces in document
interface ImageCapture
    getter videoStreamTrack
    method constructor
    method getPhotoCapabilities
    method grabFrame
    method setOptions
    method takePhoto
interface MediaSettingsRange
    getter current
    getter max
    getter min
    getter step
    method constructor
interface PhotoCapabilities
    getter brightness
    getter colorTemperature
    getter contrast
    getter exposureCompensation
    getter exposureMode
    getter fillLightMode
    getter focusMode
    getter imageHeight
    getter imageWidth
    getter iso
    getter redEyeReduction
    getter saturation
    getter sharpness
    getter whiteBalanceMode
    getter zoom
    method constructor
This is a testharness.js-based test.
PASS ImageCapture related interfaces in Origin-Trial enabled document. 
Harness: the test ran to completion.
